skill_smp_tools/requirements.txt
pytorch
pyyaml
requests
boltz[cuda]
